+++ protocol-designer/src/steplist/index.js
@@ -6,6 +6,7 @@ import selectors from './selectors'
 import * as actions from './actions'
 import * as utils from './utils'
 import {getFieldErrors, processField} from './fieldLevel'
+import {getDefaultsForStepType} from './formLevel'
 import type {FormWarning, FormWarningType} from './formLevel'
 export * from './types'
 export type {
@@ -18,6 +19,7 @@ export {
   rootReducer,
   selectors,
   getFieldErrors,
+  getDefaultsForStepType,
   processField,
   utils,
 }
